Added failure case test for destroy method, updated destroy method to pass test

+++ app/controllers/tasks_controller.rb
@@ -60,8 +60,11 @@ class TasksController < ApplicationController
 
     task = Task.find_by(id: task_id)
 
-    task.destroy
-
-    redirect_to tasks_path
+    if task
+      task.destroy
+      redirect_to tasks_path
+    else
+      redirect_to root_path, :flash => {:error => "Could not find task with id: #{task_id}"}
+    end
   end
 end
